third_party/blink/web_tests/fast/css/counters/counter-increment-001-expected.txt

This is the WebKit version of CSS 2.1 Test Suite: dynamic changes to 'counter-increment'.

PASS successfullyParsed is true

TEST COMPLETE

Before the dynamic change:
PASS internals.counterValue(document.getElementById('one')) is '1'
PASS internals.counterValue(document.getElementById('two')) is '3'
PASS internals.counterValue(document.getElementById('three')) is '4'
After the dynamic change:
PASS internals.counterValue(document.getElementById('one')) is '1'
PASS internals.counterValue(document.getElementById('two')) is '2'
PASS internals.counterValue(document.getElementById('three')) is '3'


TEST COMPLETE
